+++ tracing-mock/Cargo.toml
@@ -21,7 +21,6 @@ publish = false
 tracing = { path = "../tracing", version = "0.1.35", features = ["std", "attributes"], default-features = false }
 tracing-core = { path = "../tracing-core", version = "0.1.28", default-features = false }
 tracing-subscriber = { path = "../tracing-subscriber", version = "0.3", default-features = false, features = ["registry"], optional = true }
-tokio-test = { version = "0.4.2", optional = true }
 
 # Fix minimal-versions; tokio-test fails with otherwise acceptable 0.1.0
 tokio-stream = { version = "0.1.9", optional = true }
